// guipmc-casuistica/src/app/common/models/security/login.ts
import { TokenBody } from './tokenBody';
import { TokenInfo } from './tokenInfo';
import { UserInfo } from './userInfo';
import { Funcionalidades } from './funcionalidades';

export class Login {
  tokenBody: TokenBody;
  tokenInfo: TokenInfo;
  userInfo: UserInfo;
  funcionalidades: Funcionalidades;
}